google-apputils fixes
* Patch to make it compatible with Python 3
* Fix import path by using a runfiles alias

In the current setup, google_apputils and protobuf both have a tree like:

```
.
├── com_google_protobuf/
│   └── google/
│       └── protobuf
└── google_apputils/
    └── google/
        └── apputils
```

(Note this is a simplified view.) Here, each of the folders
`com_google_protobuf` and `google_apputils` gets added to `sys.path`.
So by default, when you have

```
from google.apputils import stopwatch
```

Python seems to do something along the lines of:

```
base = "google.apputils"
path = base.split(".")
for dir in sys.path:
    available = dir.listdir()
    if path[0] in available:
        # Look for the next bits here
```

If it does not find the desired module here, it does not continue looking
in other parts of the sys.path. So we see `google`, then we see that only
`protobuf` is available, not `apputils`.

This is fixed in later versions of Python with implicit (automatic)
namespaces. Earlier versions can use it too. But the fastest thing to do
is honestly just to include the Bazel-qualified name for apputils,
`google_apputils`. This is available via runfiles, which gets included
in `sys.path`. This way the implementation is version-agnostic.